+++ Test/Runtime/IGU/NewBehaviourScript.cs
@@ -2,6 +2,7 @@
 using System.IO;
 using UnityEngine;
 using System.Collections;
+using Cobilas.Collections;
 using Cobilas.Unity.Utility;
 using System.Collections.Generic;
 using Cobilas.Unity.Graphics.IGU;
@@ -10,106 +11,155 @@ using Cobilas.Unity.Graphics.IGU.Layouts;
 using Cobilas.Unity.Graphics.IGU.Elements;
 using Cobilas.Unity.Graphics.IGU.Interfaces;
 
-public class NewBehaviourScript : MonoBehaviour {
+public class NewBehaviourScript : MonoBehaviour, ISerializationCallbackReceiver {
 
-    public IGUStyleCustom custom;
-    public IGUStyleCustom customThumb;
-    public float value;
-    public float value2;
-    public float min;
-    public float max;
+    public Rect[] rects;
+    private bool AfterDeserialize;
 
-    public float size;
-    public bool isHoriz;
-    public Rect slider1;
-    public Rect slider2;
+    private void Start() {
+        for (int I = 0; I < 55; I++) {
+            int ide = (int)(Randomico.value * 5f);
+            ide = ide == 0 ? 1 : ide;
 
-    //public Rect horiz;
-    //public Rect horizThumb;
-    //public bool mouseThumb;
-    //public bool moveMouse;
-    //public Vector2 startPos;
-    //public Vector2 startPos2;
-    //public Vector2 startPosRect;
-
-    private void Start()
-    {
-
+            Rect temp = Rect.zero;
+            switch (ide) {
+                case 1:
+                    temp.size = new Vector2(130, 25);
+                    break;
+                case 2:
+                    temp.size = new Vector2(130, 130);
+                    break;
+                case 3:
+                    temp.size = new Vector2(50, 50);
+                    break;
+                case 4:
+                    temp.size = new Vector2(200, 200);
+                    break;
+                case 5:
+                    temp.size = new Vector2(300, 300);
+                    break;
+            }
+            temp.x = Randomico.value * (Screen.width - temp.width);
+            temp.y = Randomico.value * (Screen.height - temp.height);
+            ArrayManipulation.Add(temp, ref rects);
+        }
     }
 
+    private void OnEnable()
+    {
+        if (AfterDeserialize) {
+            StartCoroutine(WaitForEndOfFrame());
+            AfterDeserialize = false;
+        }
+    }
+
+    public int rot;
+    int batataQuente = -1;
     private void OnGUI() {
-        #region Temp code
-        //Event @event = Event.current;
-        //float mintemp;
-        //float maxtemp;
-        //if (min > max) {
-        //    maxtemp = min;
-        //    mintemp = max;
-        //} else {
-        //    mintemp = min;
-        //    maxtemp = max;
+
+
+        //int index = 0;
+        //foreach (var item in rects) {
+        //    //bool flag = false;
+        //    if (Event.current.type == EventType.Layout)
+        //        if (item.Contains(Event.current.mousePosition))
+        //            batataQuente = index;
+
+        //    Color old = GUI.color;
+        //    GUI.color = batataQuente == index ? Color.green : Color.red;
+
+        //    GUI.Box(item, $"Item:{index++}");
+
+        //    GUI.color = old;
         //}
+    }
 
-        //Rect horizThumbTemp = horizThumb;
-        //horizThumbTemp.position += horiz.position;
-        //double compri;
-        //float movx;
-        //float num1;
-        //double num2;
-        //float num3;
+    private IEnumerator WaitForEndOfFrame() {
+        while (true) {
+            yield return new WaitForEndOfFrame();
+            batataQuente = -1;
+        }
+    }
 
-        //bool isHover = horiz.Contains(@event.mousePosition);
-        //bool isHoverThumb = horizThumbTemp.Contains(@event.mousePosition);
-        //switch (@event.type) {
-        //    case EventType.MouseDown:
-        //        if (!(mouseThumb = isHoverThumb) && isHover) {
-        //            horizThumb.x = @event.mousePosition.x - horiz.x - 7f;
+    public void OnBeforeSerialize()
+    {
+        
+    }
 
-        //            num1 = startPosRect.x + (startPos2 = @event.mousePosition - startPos).x;
-        //            compri = Mathf.Abs(max - min);
-        //            num2 = horiz.width - horizThumb.width;
-        //            num3 = (float)compri / (float)num2;
-        //            num1 = (num1 * num3) / ((float)num2 * num3);
+    public void OnAfterDeserialize()
+    {
+        AfterDeserialize = true;
+    }
 
-        //            value = num1 * (float)compri + mintemp;
+    public class TDS_SeletorFisico {
+    
+    }
 
-        //            startPos = @event.mousePosition;
-        //            startPosRect = horizThumb.position;
-        //        } else {
-        //            startPos = @event.mousePosition;
-        //            startPosRect = horizThumb.position;
-        //        }
-        //        break;
-        //    case EventType.MouseDrag:
-        //        if (mouseThumb) {
-        //            num1 = startPosRect.x + (startPos2 = @event.mousePosition - startPos).x;
-        //            compri = Mathf.Abs(max - min);
-        //            num2 = horiz.width - horizThumb.width;
-        //            num3 = (float)compri / (float)num2;
-        //            num1 = (num1 * num3) / ((float)num2 * num3);
+    public class TDS_Fisica {
+        public Rect rect;
+        public bool isHotPotato;
+        public TDS_onGUI target;
 
-        //            value = num1 * (float)compri + mintemp;
-        //        }
-        //        break;
-        //    case EventType.MouseUp:
-        //        mouseThumb = false;
-        //        break;
-        //}
-        //value = Mathf.Clamp(value, mintemp, maxtemp);
+        public bool CheckHotPotato(Vector2 mousePosition) {
+            if (target != null && target.parent != null) {
+                if (target.parent is TDS_onGUI && target.parent.MyPhysic.isHotPotato) {
+                    Rect rectParent = target.parent.MyPhysic.rect;
+                    rectParent.position = Vector2.zero;
+                    float right = rect.xMax - rectParent.xMax;
+                    float left = rectParent.xMin - rect.xMin;
+                    float top = rectParent.yMin - rect.yMin;
+                    float bottom = rect.yMax - rectParent.yMax;
 
-        //compri = Mathf.Abs(max - min);
-        //movx = (horiz.width - horizThumb.width) * (float)(Math.Abs(value - mintemp) / compri);
+                    rect.width -= Mathf.Clamp(left, 0, rect.width) + Mathf.Clamp(right, 0, rect.width);
+                    rect.height -= Mathf.Clamp(top, 0, rect.height) + Mathf.Clamp(bottom, 0, rect.height);
 
-        //horizThumb.x = Mathf.Clamp(movx, 0f, horiz.width - horizThumb.width);
+                    return rect.Contains(mousePosition);
+                }
+                return false;
+            }
+            return rect.Contains(mousePosition);
+        }
+    }
 
-        //GUI.DrawTexture(horiz, Texture2D.whiteTexture, ScaleMode.StretchToFill, true, 0f, Color.green, Vector4.zero, Vector4.zero);
-        //GUI.DrawTexture(horizThumbTemp, Texture2D.whiteTexture, ScaleMode.StretchToFill, true, 0f, Color.red, Vector4.zero, Vector4.zero);
-        #endregion
+    public abstract class TDS_onGUI {
+        public abstract TDS_onGUI parent { get; set; }
+        public abstract TDS_Fisica MyPhysic { get; set; }
+        public abstract void onGUI();
+    }
 
-        int ID = GUIUtility.GetControlID(FocusType.Passive, slider1);
-        value = TDS_IGUStyle.DrawSlider(slider1, value, size, min, max, ID, custom.Style, customThumb.Style, isHoriz, false, false);
+    public class TDS_Box : TDS_onGUI {
+        public Rect rect;
+        public override TDS_onGUI parent { get; set; }
+        public override TDS_Fisica MyPhysic { get; set; }
 
-        ID = GUIUtility.GetControlID(FocusType.Passive, slider2);
-        value2 = GUI.Slider(slider2, value2, size, min, max, (GUIStyle)custom.Style, (GUIStyle)customThumb.Style, isHoriz, ID);
+        public TDS_Box()
+        {
+            MyPhysic = new TDS_Fisica();
+        }
+
+        public override void onGUI() {
+            GUI.Box(rect, "Box 1");
+        }
+    }
+
+    public class TDS_Clip : TDS_onGUI {
+        public Rect rect;
+        public Vector2 scrollView;
+        public event Action _onGUI;
+        public override TDS_onGUI parent { get; set; }
+        public override TDS_Fisica MyPhysic { get; set; }
+
+        public TDS_onGUI[] onGUIs;
+
+        public TDS_Clip()
+        {
+            MyPhysic = new TDS_Fisica();
+        }
+
+        public override void onGUI() {
+            GUI.BeginClip(rect, scrollView, Vector2.zero, false);
+            _onGUI?.Invoke();
+            GUI.EndClip();
+        }
     }
 }